--- src/modules/account/Profile/components/ProfileForm/constants/ProfileForm.constants.ts
@@ -1,3 +1,0 @@
-export const PROFILE_QUERY_KEY = {
-	FETCH_GET_PROFILE: "FETCH_GET_PROFILE",
-};
