--- src/modules/account/Profile/components/ProfileForm/constants/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./ProfileForm.constants";
-export * from "./ProfileForm.constants";
