--- src/modules/account/Profile/components/ProfileForm/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./ProfileForm";
