+++ src/modules/account/Profile/components/ProfileForm/profile-form.constants.ts
@@ -1,6 +1,10 @@
 import * as Yup from "yup";
 
-export const ProfileFormSchema = Yup.object({
+export const PROFILE_QUERY_KEY = {
+	FETCH_GET_PROFILE: "FETCH_GET_PROFILE",
+};
+
+export const PROFILE_FORM_SCHEMA = Yup.object({
 	full_name: Yup.string()
 		.matches(
 			/^([\wáàâãéèêíïóôõöúçñÁÀÂÃÉÈÍÏÓÔÕÖÚÇÑ]{2,})+\s+([\wáàâãéèêíïóôõöúçñÁÀÂÃÉÈÍÏÓÔÕÖÚÇÑ\s]{2,})+$/i,
