--- src/modules/account/Profile/components/ProfileForm/schema/index.ts
@@ -1,2 +1,0 @@
-export * as default from "./ProfileForm.schema";
-export * from "./ProfileForm.schema";
