+++ src/modules/account/Profile/services/get-profile.service.ts
@@ -1,0 +1,7 @@
+import api from "shared/services/api";
+import { IGetProfileResponse } from "modules/account/Profile/services/types";
+
+export const getProfileService = async () => {
+	const { data } = await api.get<IGetProfileResponse>("auth/account");
+	return data;
+};
