--- src/modules/account/Profile/services/types/fetchGetProfileService.types.ts
@@ -1,4 +1,0 @@
-export interface IFetchGetProfileResponse {
-	full_name: string;
-	email: string;
-}
