--- src/modules/account/Profile/services/types/fetchPutProfileService.types.ts
@@ -1,5 +1,0 @@
-export interface IFetchPutProfileService {
-	body: {
-		full_name: string;
-	};
-}
