+++ src/modules/account/Profile/services/types/put-profile.types.ts
@@ -1,0 +1,5 @@
+export interface IPutProfileService {
+	body: {
+		full_name: string;
+	};
+}
